+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.7.7 (2021-04-06)
 - add the new F34 i3 spin
+- shorten mate_compiz to mate
 
 ## 0.7.6 (2021-01-21)
 - improve help text for releases related for Beta and RCs (#1)